"Finish" Project
* Remake main.ejs login page
* Implement user functionality and front-end
* Remake routes to check a user has the permission to execute an action
* Update JSDoc
* Add more utility functions in DB.js
* Update DBScripts.js
* Fix bugs in admin.ejs

+++ app/app.js
@@ -11,6 +11,16 @@ function getPath(IsAdmin) {
     return "/user";
 }
 
+async function hasPermission(userId, itemId) {
+    let user = await DBM.getUserById(userId);
+    console.log(user[0].IsAdmin);
+    if(user[0].IsAdmin)
+        return 1;
+    if(userId==itemId)
+        return 1;
+    return 0;
+}
+
 app.use(session({
     name: sessionConfig.name,
     resave: false,
@@ -67,13 +77,21 @@ app.get("/admin", redirectLogin, async (req, res)=>{
         tasks: await DBM.getAllTasks(),
         worklogs: await DBM.getAllWorklogs(),
         teams: await DBM.getAllTeams(),
-        realUsersCount: await DBM.getRealUserCount()
+        realUsersCount: await DBM.getRealUserCount(),
     };
     res.render("../views/admin.ejs", { data });
 });
 
-app.get("/user", redirectLogin, (req, res)=>{
-    res.render("../views/user.ejs");
+app.get("/user", redirectLogin, async (req, res)=>{
+    let user = await DBM. getPublicUserDataById(req.session.userId);
+    let data = await DBM.getRelevantDataByUserId(req.session.userId);
+    data.users=await DBM.getAllUsers();
+    data.firstName = user.FirstName;
+    data.lastName = user.LastName;
+    data.userId = req.session.userId;
+    data.allTeams = await DBM.getAllTeams();
+    data.allTasks = await DBM.getAllTasks();
+    res.render("../views/user.ejs", { data } );
 });
 
 // app.get("*", (req,res)=>{
@@ -110,15 +128,12 @@ app.post("/logout", (req, res)=>{
     });
 });
 
-// app.post("/delete-project", async (req,res)=>{
-//     const { id } = req.body;
-//     await DBM.deleteProject(id)
-//     res.redirect("/admin");
-// });
-
 app.post("/delete-project/:id", async(req,res) =>{
     const { id } = req.params;
-    await DBM.deleteProject(id);
+    const project = await DBM.getProjectById(id);
+    if(await hasPermission(req.session.userId,project.CreatorId))
+        await DBM.deleteProject(id);
+    res.sendStatus(204);
 });
 
 app.get("/get-users-teams/:id", async(req,res)=>{
@@ -131,21 +146,30 @@ app.get("/get-users-teams/:id", async(req,res)=>{
 app.post("/create-project/:title/:description", async (req,res)=>{
     const project = { title: req.params.title, description: req.params.description, creatorId: req.session.userId };
     await DBM.createProject(project);
+    let user = await DBM.getUserById(req.session.userId);
+    console.log(user);
+    if(user[0].IsAdmin) {
+        res.redirect("/admin");
+    } else {
+        res.redirect("/user");
+    }
 });
 
 app.post("/update-project", async(req,res)=>{
     const newProject = req.body;
+    console.log(newProject);
     newProject.teams = newProject.teams.split(",");
     newProject.tasks = newProject.tasks.split(",");
-    
-    await DBM.updateProject(newProject.projectId,newProject,req.session.userId);
-    console.log(newProject.teams);
-    if(newProject.teams)
-        await DBM.addTeamsToProject(newProject.teams, newProject.projectId);
-    if(newProject.task)
-        await DBM.addTasksToProject(newProject.tasks, newProject.projectId);
-
-    res.redirect("/admin");
+    let project=await DBM.getProjectById(newProject.projectId);
+    if(await hasPermission(req.session.userId,project.CreatorId)) {
+        await DBM.updateProject(newProject.projectId,newProject,req.session.userId);
+        console.log(project);
+        if(newProject.teams)
+            await DBM.addTeamsToProject(newProject.teams, newProject.projectId);
+        if(newProject.tasks)
+            await DBM.addTasksToProject(newProject.tasks, newProject.projectId);
+    }
+    res.redirect("/user");
 });
 
 app.get("/get-team-users/:teamId", async(req,res)=>{
@@ -173,7 +197,12 @@ app.post("/create-task", async (req,res)=>{
 
 app.post("/delete-task/:id", async (req,res)=>{
     const id = req.params.id;
-    await DBM.deleteTask(id);
+    let task = await DBM.getTaskById(id);
+    console.log(task.CreatorId)
+    console.log(req.session.userId);
+    if(await hasPermission(req.session.userId, task.CreatorId)) {
+        await DBM.deleteTask(id);
+    }
     res.sendStatus(204);
 })
 
@@ -228,7 +257,10 @@ app.post("/update-team",async(req,res)=>{
 app.post("/update-task", async(req,res)=>{
     let newTask = req.body;
     console.log(newTask);
-    await DBM.updateTask(newTask.taskId,newTask,req.session.userId);
+    const task = await DBM.getTaskById(newTask.taskId);
+    if(await hasPermission(req.session.userId,task.CreatorId)) {
+        await DBM.updateTask(newTask.taskId,newTask,req.session.userId);
+    }
     res.redirect("/admin");
 })
 
@@ -242,15 +274,20 @@ app.post("/create-worklog",async(req,res)=>{
 
 app.post("/delete-worklog/:id",async(req,res)=>{
     let id = req.params.id;
-    await DBM.deleteWorklog(id);
+    let log = await DBM.getWorklogById(id);
+    if(await hasPermission(req.session.userId,log.CreatorId))
+        await DBM.deleteWorklog(id);
     res.sendStatus(204);
 })
 
 app.post("/update-worklog", async (req,res)=>{
     let worklog = req.body;
     worklog.worklogId=parseInt(worklog.worklogId);
-    worklog.time=parseInt(worklog.time);
-    console.log(worklog);
-    await DBM.updateWorklog(worklog.worklogId,worklog);
+    let log = await DBM.getWorklogById(worklog.worklogId);
+    if(await hasPermission(req.session.userId),log.CreatorId) {
+        worklog.time=parseInt(worklog.time);
+        console.log(worklog);
+        await DBM.updateWorklog(worklog.worklogId,worklog);
+    }
     res.redirect("/admin");
 })
